create dev and devServer webpack profiles
Instead of producing an external source map, `yarn run build-dev`
creates build/oligrapher-dev.js without minification and an
inline source map.

`yarn run dev-server` starts th webpack-dev-serve, which is what
used to be started by `yarn run dev`
