+++ babel.config.js
@@ -1,8 +1,10 @@
 module.exports = function (api) {
+  // These are run in reserve order. @babel/preset-react is run first.
   const presets = [
-    "@babel/preset-env",
-    ["@babel/preset-react", { "modules": false } ]
+    [ "@babel/preset-env",   { useBuiltIns: "usage", corejs: 3 } ],
+    [ "@babel/preset-react", {} ]
   ];
+
   const plugins = [];
 
   api.cache.never();
