+++ webpack.prod.js
@@ -1,12 +1,10 @@
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const merge = require('webpack-merge');
 const commonConfig = require('./webpack.common.js');
 
 const productionConfig = {
   mode: 'production',
   optimization: {
-    minimize: true,
-    minimizer: [new UglifyJsPlugin()],
+    minimize: true
   }
 };
